Allow to use the Shalon URL protocol handler for P-Services
As the `PrivacyService.KeyValueStorage` did use separate URL sessions
registering the `ShalonURLProtocol` as a URL handler class within the
application would not work. The globally registered URL handlers are
only active for the shared URL session. Therefore the URL handler is
registered for the ephemeral sessions used for requesting content from
the P-Service. If the base URL of the P-Service is configured as a
Shalon-URL, then all requests will be proxied.