Add default port for proxies in the Shalon URL protocol
This also simplified the code quite a bit. This also led me to detect
another bug. If to few proxies were specified, the correct error was not
thrown, as the check for it was off by one. The tests succeeded, as they
only check that an error is thrown, but do not validate the error
itself. The actual error thrown was an invalid proxy due to the missing
port. The tests now check for the correct error type as well.

This resolves issue #7.